+++ src/main/java/com/example/JJShop/service/ItemServiceImpl.java
@@ -2,6 +2,7 @@ package com.example.JJShop.service;
 
 import com.example.JJShop.exceptions.AlreadyCreatedException;
 import com.example.JJShop.exceptions.NotFoundException;
+import com.example.JJShop.model.Category;
 import com.example.JJShop.model.Item;
 import com.example.JJShop.model.enums.ErrorMessages;
 import com.example.JJShop.repository.ItemRepository;
@@ -12,7 +13,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Service
-public class ItemServiceImpl implements ItemService  {
+public class ItemServiceImpl implements ItemService {
     private final ItemRepository itemRepository;
 
     public ItemServiceImpl(ItemRepository itemRepository) {
@@ -22,7 +23,7 @@ public class ItemServiceImpl implements ItemService  {
     @Override
     public Item createItem(Item item) {
         Optional<Item> itemFindByItemName = itemRepository.findByItemName(item.getItemName());
-        if (itemFindByItemName.isPresent()){
+        if (itemFindByItemName.isPresent()) {
             throw new AlreadyCreatedException(ErrorMessages.NAME_EXISTS.getMessage());
         }
         return itemRepository.save(item);
@@ -34,24 +35,24 @@ public class ItemServiceImpl implements ItemService  {
         if (dbItem.isEmpty()) {
             throw new NotFoundException(ErrorMessages.NAME_NOT_FOUND.getMessage());
         }
-        Optional<Item> itemFindByItemName = itemRepository.findByItemNameAndItemIdNot(updatedItem.getItemName(), id);
-        if (itemFindByItemName.isPresent()){
-            throw new AlreadyCreatedException(ErrorMessages.NAME_EXISTS.getMessage());
-        }
-
-        dbItem.get().setItemName(updatedItem.getItemName());
-        dbItem.get().setItemDescription(updatedItem.getItemDescription());
-        dbItem.get().setItemPrice(updatedItem.getItemPrice());
-        dbItem.get().setIsItemAvailable(updatedItem.getIsItemAvailable());
-        dbItem.get().setItemStock(updatedItem.getItemStock());
-        dbItem.get().setCategory(updatedItem.getCategory());
+            if (dbItem.get().getItemName().equals(updatedItem.getItemName())) {
+                throw new AlreadyCreatedException(ErrorMessages.NAME_EXISTS.getMessage());
+            }
+            else {
+                dbItem.get().setItemName(updatedItem.getItemName());
+                dbItem.get().setItemDescription(updatedItem.getItemDescription());
+                dbItem.get().setItemPrice(updatedItem.getItemPrice());
+                dbItem.get().setIsItemAvailable(updatedItem.getIsItemAvailable());
+                dbItem.get().setItemStock(updatedItem.getItemStock());
+                dbItem.get().setCategory(updatedItem.getCategory());
+            }
         return itemRepository.save(dbItem.get());
     }
 
     @Override
     public Item getItemById(Long id) {
         Optional<Item> item = itemRepository.findById(id);
-        if (item.isEmpty()){
+        if (item.isEmpty()) {
             throw new NotFoundException(ErrorMessages.NAME_NOT_FOUND.getMessage());
         }
         return item.get();
@@ -59,12 +60,21 @@ public class ItemServiceImpl implements ItemService  {
 
     @Override
     public List<Item> findAllItems() {
-        return (List<Item>) itemRepository.findAll();
+        List<Item> items = (List<Item>) itemRepository.findAll();
+        if (items.isEmpty()) {
+            throw new NotFoundException(ErrorMessages.NAME_NOT_FOUND.getMessage());
+        }
+        return items;
     }
 
     @Override
     public void deleteItemById(Long id) {
-        itemRepository.deleteById(id);
+        Optional<Item> item = itemRepository.findById(id);
+        if (item.isEmpty()) {
+            throw new NotFoundException(ErrorMessages.NAME_NOT_FOUND.getMessage());
+        } else {
+            itemRepository.deleteById(id);
+        }
     }
 
 }
